requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

sqlalchemy==2.0.29

python-dotenv==1.0.1

groq==0.11.0

sentence-transformers==2.6.1
numpy==1.26.4

langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-groq==0.1.4

chromadb==0.5.0